Refactor JNI tests
