+++ sherpa-onnx/jni/jni.cc
@@ -4,6 +4,9 @@
 //                2022       Pingfeng Luo
 //                2023       Zhaoming
 
+#include <fstream>
+
+#include "sherpa-onnx/csrc/macros.h"
 #include "sherpa-onnx/csrc/wave-reader.h"
 #include "sherpa-onnx/csrc/wave-writer.h"
 #include "sherpa-onnx/jni/common.h"
